Start On Password Module
Password Module:
This is a mini-game module that the player needs to enter a password in to pass the module.
Implemented the basic functionality of the module without art assets.
To Do's Still: art, confirm instead of auto confirm, randomise the password

Technical breakdown:
Added ModuleButtton, this can be attached to anything and will trigger an event when it has been pressed, with the value of the button
Added PasswordModule - compares string entered to a hardcoded password
Added IModule - will become the interface used for the modules to keep minigames in a modular pattern.